+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - front-end: dispatch week/day management in js files
 - back-end refactoring: all model attributes in english
+- Selection of multiple weeks possible with the solve board
 
 ### TBD
 - right permissions in TTapp/views